requirements.txt
Babel==2.3.4
defusedxml==0.4.1
Django==1.9.12
django-appconf==1.0.2
django-compressor==2.1
django-extra-views==0.6.4
django-haystack==2.5.1
django-modeltranslation==0.12
django-oscar==1.3
django-tables2==1.0.7
django-treebeard==4.1.0
django-widget-tweaks==1.4.1
factory-boy==2.7.0
fake-factory==0.7.2
mock==2.0.0
oauthlib==2.0.1
pbr==1.10.0
phonenumbers==7.7.5
Pillow==3.4.2
psycopg2==2.6.2
purl==1.3
pycountry==16.11.8
PyJWT==1.4.2
pysolr==3.6.0
python-dateutil==2.6.0
python-social-auth==0.2.21
python3-openid==3.0.10
pytz==2016.7
rcssmin==1.0.6
requests==2.12.1
requests-oauthlib==0.7.0
rjsmin==1.0.12
six==1.10.0
sorl-thumbnail==12.4a1
Unidecode==0.4.19
vk-api==8.0
